don't import runtime dependencies in addon definition
